fix(metrics): corrigir bugs nos endpoints de métricas após testes e2e
- DailyCountProjection/DailyDurationProjection: java.sql.Date → LocalDate
  (driver PostgreSQL moderno retorna LocalDate para DATE(col); Spring Data
  não conseguia projetar de volta para sql.Date — causava 500 em /overview e /flow)
- OverviewMetricsService/FlowMetricsService: remover .toLocalDate() redundante
- CollaborationMetricsResponse.TeamMetrics: Long commits/prsMerged → Double
  (divisão inteira 360/444=0 inflava teamAverage zerado — bug numérico)
- PullRequestCommitRepository: cycle time usa GREATEST(MIN(commit_date), pr.created_at)
  (sem GREATEST, commits antigos no branch inflavam média para ~250 dias)
- IssueRepository: lead time adiciona AND created_at >= :from
  (issues abertas há anos e fechadas no período inflavam média para ~557 dias)

+++ src/main/java/br/com/tcc/github_poc/metrics/dto/CollaborationMetricsResponse.java
@@ -14,5 +14,5 @@ public record CollaborationMetricsResponse(
 
     public record ComparisonMetrics(TeamMetrics individual, TeamMetrics teamAverage) {}
 
-    public record TeamMetrics(Long commits, Long prsMerged, Double tcm) {}
+    public record TeamMetrics(Double commits, Double prsMerged, Double tcm) {}
 }
